refactor: sponsor component to create a server island for sponsors, currently the sponsors are load only when build - No real-fetch, i do not implement this because it has deploy considerations so i prefer to leave it ready to implement
