+++ src/constants/featured-sponsors.ts
