+++ src/sections/footer.astro
@@ -1,6 +1,7 @@
 ---
 import { fetchSponsors } from "@/utils/fetch-sponsors";
 const { sponsors } = await fetchSponsors();
+import Sponsor from "../components/ui/sponsor.astro";
 ---
 
 <footer class="text-zinc-600 space-y-4 pb-8">
@@ -17,23 +18,7 @@ const { sponsors } = await fetchSponsors();
   <div class="flex items-center -space-x-1.5">
     {
       sponsors.map((s, i) => (
-        <a
-          href={`https://github.com/${s.login}`}
-          target="_blank"
-          rel="noopener noreferrer"
-          class="sponsor-avatar relative hover:z-10"
-          style={`--i: ${i}`}
-        >
-          <img
-            src={`https://unavatar.io/github/${s.login}`}
-            alt={s.login}
-            width="32"
-            height="32"
-            class="size-8 rounded-full border-2 border-zinc-900 shadow-sm shadow-black/30"
-            loading="lazy"
-          />
-          <span class="sponsor-tip">@{s.login}</span>
-        </a>
+        <Sponsor s={s} i={i} />
       ))
     }
